+++ include/objectGroups.h
@@ -3,7 +3,7 @@
  
  * File:   include/objectGroups.h
  *
- * Copyright (c) 2004-2018 by Stuart Ansell
+ * Copyright (c) 2004-2019 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -81,7 +81,7 @@ class objectGroups
 
   int getNextCell(const int) const;
   
-  int cell(const std::string&,const int = 10000);
+  int cell(const std::string&,const size_t = 10000);
   
   std::string inRange(const int) const;
   bool hasCell(const std::string&,const int) const;
